% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckFunctions.R
\name{checkChemical}
\alias{checkChemical}
\title{Check chemicals are supported for analysis}
\usage{
checkChemical(cname)
}
\arguments{
\item{cname}{List of chemical characters.}
}
\value{
Tibble data frame with chemicals and if they are supported or not.
}
\description{
Check chemicals are supported for analysis
}
\examples{
checkChemical(list("Ethanol","Iodine","Zinc"))
}
