% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckFunctions.R
\name{checkHPO}
\alias{checkHPO}
\title{Check HPO terms are supported for analysis}
\usage{
checkHPO(HPO)
}
\arguments{
\item{HPO}{List of Human Phenotype Ontology terms characters.}
}
\value{
Tibble data frame with HPO terms and if they are supported or not.
}
\description{
Check HPO terms are supported for analysis
}
\examples{
checkHPO(list("Rickets","Breast carcinoma","Preeclampsia"))
}
