% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{chem2gene}
\alias{chem2gene}
\title{Chemical and gene interactions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1575035 rows and 11 columns.}
\usage{
chem2gene
}
\description{
Dataset of chemicals and their gene interactions. Original data set downloaded on 11-Feb-2019. Version Feb-05-2019.
Dataset has been filtered to contain only the organisms Homo sapiens, Mus musculus, Rattus norvegicus,
Rattus rattus, Danio rerio, Macaca fascicularis, Macaca mulatta and human gene IDs. Genes were checked that
they belonged to humans using the Homo sapien gene info from NCBI (downloaded 19-March-2019).
For gene IDs not belonging to Homo sapiens, gene IDs were checked that they belonged to their respective organism
using All Mammalia gene info from NCBI (15-March-2019) and Danio rerio gene info from NCBI (15-March-2019).
All orthologs for these gene IDs were found using NCBI Homologene (version 06-May-2014) and outdated ortholog gene IDs
were updated to the latest gene ID using NCBI gene_history (15-March-2019) and updated gene symbols were taken from
Homo sapien gene info.
}
\details{
@format A data frame of 1,575,035 rows and 11 columns.



@source \url{http://ctdbase.org/downloads/#cg}
}
\keyword{datasets}
