% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalFunctions.R
\name{perfFishTestChemMultiple}
\alias{perfFishTestChemMultiple}
\title{Enriched HPO terms for multiple chemicals}
\usage{
perfFishTestChemMultiple(cname, enrich_1S = TRUE)
}
\arguments{
\item{cname}{List of chemical characters.}

\item{enrich_1S}{Enrichment one-sided. Default set to true for enrichment, where the one-sided Fisher's exact is set to alternative = "greater."
If enrich_1S = FALSE, Fisher's exact is set to alternative = "two.sided" for two-sided.}
}
\value{
Tibble data frame with enrichment.
}
\description{
Enriched HPO terms for multiple chemicals
}
\note{
Enrichment for multiple chemicals is a simplistic exploratory function, where we take an additive approach.
Chemical names and synonyms used are dependent on the chemical names within CTD.
}
\examples{
perfFishTestChemMultiple(list("bisphenol A","bisphenol F","bisphenol B"))
}
