% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalFunctions.R
\name{perfFishTestChemSingle}
\alias{perfFishTestChemSingle}
\title{Enriched HPO terms for a single chemical's associated gene list}
\usage{
perfFishTestChemSingle(cname, enrich_1S = TRUE)
}
\arguments{
\item{cname}{Chemical character}

\item{enrich_1S}{Enrichment one-sided. Automatically set to true for enrichment, where the one-sided Fisher's exact is set to alternative = 'greater.'
If enrich_1S = FALSE, Fisher's exact is set to alternative = "two.sided" for two-sided.}
}
\value{
Dataframe with enrichment.
}
\description{
Enriched HPO terms for a single chemical's associated gene list
}
\note{
Chemical names and synonyms used are dependent on the chemical names within CTD.
}
\examples{
perfFishTestChemSingle("Iodine")
}
