% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPOFunctions.R
\name{perfFishTestHPOMultiple}
\alias{perfFishTestHPOMultiple}
\title{Chemical enrichment for multiple phenotype terms}
\usage{
perfFishTestHPOMultiple(HPO, enrich_1S = TRUE)
}
\arguments{
\item{HPO}{List of Human Phenotype Ontology terms characters}

\item{enrich_1S}{Enrichment one-sided. Default set to true for enrichment, where the one-sided Fisher's exact is set to alternative = "greater."
If enrich_1S = FALSE, Fisher's exact is set to alternative = "two.sided" for two.sided.}
}
\value{
Tibble data frame with enrichment.
}
\description{
Chemical enrichment for multiple phenotype terms
}
\note{
Chemical enrichment for multiple HPO terms is a simplistic exploratory function, where we take an additive approach.
HPO terms used should be phenotypic abnormality terms within the HPO release 2019-02-12.
}
\examples{
perfFishTestHPOMultiple(list("Increased circulating ACTH level",
"Androgen insufficiency",
"Decreased circulating aldosterone level"))
}
