% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{pheno2gene}
\alias{pheno2gene}
\title{Phenotype and gene association}
\format{An object of class \code{data.frame} with 472515 rows and 4 columns.}
\usage{
pheno2gene
}
\description{
Dataset of phenotype and their gene associations. Original dataset downloaded on 28-Feb-2019. Version 2019-02-12.
Dataset has been filtered to only contain phenotypic abnormality terms. Generation of this dataset used the Human Phenotype Ontology OBO file (2019-02-12 version)
and the R packages ontologyIndex, tibble, magrittr and dplyr.
}
\details{
@format A data frame of 472,515 rows and 4 columns.



@source \url{https://hpo.jax.org/app/download/annotation}
}
\keyword{datasets}
