% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{sym_tbl}
\alias{sym_tbl}
\title{Chemical vocabulary}
\format{An object of class \code{data.frame} with 63819 rows and 2 columns.}
\usage{
sym_tbl
}
\description{
Dataset of chemicals and their synonyms. Data has been modified from original data to contain synonyms only for chemicals with gene interactions.
Duplicated synonyms for tetrachlorodibenzodioxin and polychlorinated dibenzodioxins were removed from polychlorinated dibenzodioxins,
synonyms have been changed to lowercase and duplicate synonyms have been removed. Downloaded on 11-Feb-2019. Version Feb-05-2019. The R packages dplyr, tidyr,
stringr, magrittr and tibble were used in the generation of this dataset.
}
\details{
@format A data frame of 63,819 rows and 2 columns.



@source \url{http://ctdbase.org/downloads/#allchems}
}
\keyword{datasets}
