% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunction.R
\name{visEnrich}
\alias{visEnrich}
\title{Enrichment visualisation}
\usage{
visEnrich(x)
}
\arguments{
\item{x}{Dataframe of enrichment data}
}
\value{
Shiny app.
}
\description{
Enrichment visualisation
}
\examples{
\dontrun{
x <- perfFishTestHPOSingle("Iodine")
visEnrich(x)}




}
