% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPOFunctions.R
\name{perfFishTestHPOMultiple}
\alias{perfFishTestHPOMultiple}
\title{Chemical enrichment for multiple phenotype terms}
\usage{
perfFishTestHPOMultiple(HPO, enrich_1S = TRUE)
}
\arguments{
\item{HPO}{List of Human Phenotype Ontology terms}

\item{enrich_1S}{Enrichment one-sided. Automatically set to true for enrichment, where the one-sided Fisher's exact is set to alternative = 'greater.'
If enrich_1S = FALSE, Fisher's exact is set to alternative = 'two.sided' for two.sided.}
}
\value{
Dataframe with enrichment.
}
\description{
Chemical enrichment for multiple phenotype terms
}
\note{
Chemical enrichment for multiple HPO terms is a simplistic exploratory function, where we take an additive approach.
HPO terms used should be phenotypic abnormality terms within the HPO release 2019-02-12.
}
\examples{
\dontrun{
x <- list("HPO term1","HPO term2","HPO term3")
perfFishTestHPOMultiple(x)
}
}
