% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{sym_tbl}
\alias{sym_tbl}
\title{Chemical vocabularly}
\format{An object of class \code{data.frame} with 63818 rows and 2 columns.}
\usage{
sym_tbl
}
\description{
Dataset of chemicals and their synonyms. Data has been modified from original data, to contain synonyms only for chemicals with gene interactions,
synonyms have been changed to lower-case and duplicate synonyms have been removed. Downloaded on 11-Feb-2019. Version Feb-05-2019.
}
\details{
@format A data frame of 63,818 rows and 2 columns.



@source \url{http://ctdbase.org/downloads/#allchems}
}
\keyword{datasets}
